Extract lower-level edit utility from autofix module
