+++ crates/ruff/src/rules/flake8_pie/rules/duplicate_class_field_definition.rs
@@ -9,7 +9,7 @@ use ruff_macros::{derive_message_formats, violation};
 
 use ruff_python_ast::types::RefEquality;
 
-use crate::autofix::actions::delete_stmt;
+use crate::autofix::edits::delete_stmt;
 use crate::checkers::ast::Checker;
 use crate::registry::AsRule;
