+++ crates/ruff/src/rules/flake8_pytest_style/rules/fixture.rs
@@ -15,7 +15,7 @@ use ruff_python_ast::{helpers, visitor};
 use ruff_python_semantic::analyze::visibility::is_abstract;
 use ruff_python_semantic::model::SemanticModel;
 
-use crate::autofix::actions::remove_argument;
+use crate::autofix::edits::remove_argument;
 use crate::checkers::ast::Checker;
 use crate::registry::{AsRule, Rule};
