+++ crates/ruff/src/rules/pyflakes/rules/unused_variable.rs
@@ -11,7 +11,7 @@ use ruff_python_ast::source_code::Locator;
 use ruff_python_ast::types::RefEquality;
 use ruff_python_semantic::scope::{ScopeId, ScopeKind};
 
-use crate::autofix::actions::delete_stmt;
+use crate::autofix::edits::delete_stmt;
 use crate::checkers::ast::Checker;
 use crate::registry::AsRule;
